chore: Integrate Ruff Linter and Resolve Code Issues
This commit introduces the integration of the Ruff linter into the project and addresses linting issues that were detected. The changes include modifications to the following files:

- cache/language_redis_manager.py
- cache/redis_utils.py
- cache/users_redis_manager.py
- config.py
- conftest.py
- db/db_session.py
- db/models/user_model.py
- db/operations/users_operations.py
- tgbot/__main__.py
- tgbot/handlers/common_handles/__init__.py
- tgbot/handlers/common_handles/start_handler.py
- tgbot/handlers/common_handles/user_group_status.py
- tgbot/keyboards/app_buttons.py
- tgbot/middlewares/apscheduler_middleware.py
- tgbot/middlewares/button_middleware.py
- tgbot/middlewares/chat_member_middleware.py
- tgbot/middlewares/db_middleware.py
- tgbot/middlewares/redis_middleware.py
- tgbot/middlewares/throttling_middleware.py
- tgbot/middlewares/translation_middleware.py
- tgbot/schedule/schedule_adjustment.py
- tgbot/template_engine/jinja_engine.py
- tgbot/utils/before_bot_start.py
- tgbot/utils/bot_commands.py

The linting issues identified, including rule violations E501, UP007, UP035, UP006, D103, and D212, have been resolved to ensure that the codebase adheres to linting standards. This integration and code cleanup enhance code quality and maintainability.

This commit message serves to document the introduction of the Ruff linter and the improvements made to maintain code consistency and quality.