ci: Update Docker image build workflow
This commit updates the workflow for building and pushing the Docker image to Docker Hub. The changes include using the latest version of the GitHub Actions `docker/setup-buildx-action` and `docker/build-push-action`. It also sets up the necessary authentication for Docker Hub using secrets.

The workflow now builds and pushes the Docker image for the "tracker-tgbot" repository with the tag "0.0.3" to Docker Hub.